remove the null bundle test since it does not really solve the problem.
To actually solve this extreme corner case we would need check if the ServiceReference can still be obtained from the service registry.  This is the only way to tell that the service is in the process of being unregistered as the service is removed from the service registry before the UNREGISTERING event is fired.

For example:

if (context.getServiceReferences(null, "(service.id="+reference.getProperty(Constants.SERVICE_ID)+")") == null) {
  // we have the corner case
}

but this is gross overkill for something that will in all likelyhood never happen in the real world.

And this test must also be done in the synchronized (this) surrouding the 2 places where added.add() is called.